Refactoring the remainder of Inverse Trigonometry
