+++ src/Samsara/Fermat/Core/Types/Traits/Trigonometry/InverseTrigonometryHelpersTrait.php
@@ -2,13 +2,16 @@
 
 namespace Samsara\Fermat\Core\Types\Traits\Trigonometry;
 
+use Samsara\Exceptions\SystemError\LogicalError\IncompatibleObjectState;
 use Samsara\Exceptions\SystemError\PlatformError\MissingPackage;
 use Samsara\Exceptions\UsageError\IntegrityConstraint;
+use Samsara\Fermat\Core\Enums\CalcOperation;
 use Samsara\Fermat\Core\Numbers;
 use Samsara\Fermat\Core\Provider\SequenceProvider;
 use Samsara\Fermat\Core\Provider\SeriesProvider;
 use Samsara\Fermat\Core\Types\Base\Interfaces\Callables\ContinuedFractionTermInterface;
 use Samsara\Fermat\Core\Values\ImmutableDecimal;
+use Samsara\Fermat\Core\Values\MutableDecimal;
 
 /**
  * @package Samsara\Fermat\Core
@@ -151,6 +154,238 @@ trait InverseTrigonometryHelpersTrait
 
         return $x->multiply($one->subtract($x2)->sqrt($scale))->divide($answer, $scale);
 
+    }
+
+    /**
+     * @param CalcOperation $operation
+     * @param int|null $scale
+     * @param string $oneInputVal
+     * @param string $negOneInputVal
+     * @param bool $round
+     * @return static|ImmutableDecimal|MutableDecimal
+     * @throws IntegrityConstraint
+     */
+    protected function helperArcsecArccscSimple(
+        CalcOperation $operation,
+        ?int $scale,
+        string $oneInputVal,
+        string $negOneInputVal,
+        bool $round
+    ): static|ImmutableDecimal|MutableDecimal
+    {
+        $abs = $this instanceof ImmutableDecimal ? $this->abs() : new ImmutableDecimal($this->absValue());
+        if ($abs->isLessThan(1)) {
+            throw new IntegrityConstraint(
+                'The input for arcsec and arccsc must have an absolute value greater than or equal to 1',
+                'Only calculate arcsec and arccsc for values greater than 1',
+                'The arcsec and arccsc function only have real values for inputs which have an absolute value greater than or equal to 1'
+            );
+        }
+
+        $finalScale = $scale ?? $this->getScale();
+
+        if ($this->isEqual(1)) {
+            $answer = $oneInputVal;
+        } elseif ($this->isEqual(-1)) {
+            $answer = $negOneInputVal;
+        } else {
+            $answer = match ($operation) {
+                CalcOperation::ArcSec => $this->arcsecSelector($scale),
+                CalcOperation::ArcCsc => $this->arccscSelector($scale),
+            };
+        }
+
+        if ($round) {
+            $result = $this->setValue($answer)->roundToScale($finalScale);
+        } else {
+            $result = $this->setValue($answer)->truncateToScale($finalScale);
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param CalcOperation $operation
+     * @param int|null $scale
+     * @param bool $round
+     * @return ImmutableDecimal|MutableDecimal|static
+     * @throws IntegrityConstraint
+     */
+    protected function helperArcsinArccosSimple(
+        CalcOperation $operation,
+        ?int $scale = null,
+        bool $round = true
+    ): ImmutableDecimal|MutableDecimal|static
+    {
+        $abs = $this instanceof ImmutableDecimal ? $this->abs() : new ImmutableDecimal($this->absValue());
+        if ($abs->isGreaterThan(1)) {
+            throw new IntegrityConstraint(
+                'The input for arcsin and arccos must have an absolute value of 1 or smaller',
+                'Only calculate arcsin and arccos for values of 1 or smaller',
+                'The arcsin and arccos function only have real values for inputs which have an absolute value of 1 or smaller'
+            );
+        }
+
+        return $this->helperArcBasicSimple($operation, $scale, $round);
+    }
+
+    /**
+     * @param CalcOperation $operation
+     * @param int|null $scale
+     * @param bool $round
+     * @return static|ImmutableDecimal|MutableDecimal
+     * @throws IntegrityConstraint
+     */
+    protected function helperArcBasicSimple(
+        CalcOperation $operation,
+        ?int $scale,
+        bool $round
+    ): static|ImmutableDecimal|MutableDecimal
+    {
+        $finalScale = $scale ?? $this->getScale();
+
+        $answer = match ($operation) {
+            CalcOperation::ArcTan => $this->arctanSelector($scale),
+            CalcOperation::ArcCot => $this->arccotSelector($scale),
+            CalcOperation::ArcSin => $this->arcsinSelector($scale),
+            CalcOperation::ArcCos => $this->arccosSelector($scale),
+        };
+
+        if ($round) {
+            $result = $this->setValue($answer)->roundToScale($finalScale);
+        } else {
+            $result = $this->setValue($answer)->truncateToScale($finalScale);
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param ImmutableDecimal $zeroTerm
+     * @param CalcOperation $operation
+     * @param int $scale
+     * @return string
+     * @throws IncompatibleObjectState
+     * @throws IntegrityConstraint
+     * @throws MissingPackage
+     */
+    protected function helperArcsecArccscScale(ImmutableDecimal $zeroTerm, CalcOperation $operation, int $scale): string
+    {
+        $one = Numbers::makeOne($scale);
+        $z = Numbers::makeOrDont(Numbers::IMMUTABLE, $this, $scale);
+        $aPart = $this->helperArcsecArccscAPart($one, $z, $scale);
+
+        $bPart = $this->helperArcsecArccscBPart($scale, $zeroTerm);
+
+        $answer = SeriesProvider::generalizedContinuedFraction(
+            $aPart,
+            $bPart,
+            $scale * 2,
+            $scale,
+            match ($operation) {
+                CalcOperation::ArcCsc => SeriesProvider::SUM_MODE_ALT_FIRST_ADD,
+                CalcOperation::ArcSec => SeriesProvider::SUM_MODE_SUB,
+            }
+        );
+
+        return $answer->getAsBaseTenRealNumber();
+    }
+
+    /**
+     * @param ImmutableDecimal $one
+     * @param ImmutableDecimal $z
+     * @param int|null $intScale
+     * @return callable|ContinuedFractionTermInterface
+     * @throws IncompatibleObjectState
+     * @throws IntegrityConstraint
+     */
+    protected function helperArcsecArccscAPart(
+        ImmutableDecimal $one,
+        ImmutableDecimal $z,
+        ?int $intScale): callable|ContinuedFractionTermInterface
+    {
+        $oneDivZSquared = $one->divide($z->pow(2));
+        $oneDivZ = $one->divide($z, $intScale);
+
+        /**
+         * @package Samsara\Fermat\Core
+         */
+        return new class($oneDivZSquared, $oneDivZ, $one) implements ContinuedFractionTermInterface {
+            private ImmutableDecimal $oneDivZSq;
+            private ImmutableDecimal $oneDivZ;
+            private ImmutableDecimal $one;
+
+            /**
+             * @param ImmutableDecimal $oneDivZSq
+             * @param ImmutableDecimal $oneDivZ
+             * @param ImmutableDecimal $one
+             */
+            public function __construct(ImmutableDecimal $oneDivZSq, ImmutableDecimal $oneDivZ, ImmutableDecimal $one)
+            {
+                $this->oneDivZSq = $oneDivZSq;
+                $this->oneDivZ = $oneDivZ;
+                $this->one = $one;
+            }
+
+            /**
+             * @param int $n
+             * @return ImmutableDecimal
+             */
+            public function __invoke(int $n): ImmutableDecimal
+            {
+                if ($n == 1) {
+                    return $this->oneDivZ->multiply($this->one->subtract($this->oneDivZSq)->sqrt());
+                }
+
+                $subterm = floor($n / 2);
+
+                return $this->oneDivZSq->multiply(
+                    2 * $subterm - 1
+                )->multiply(2 * $subterm);
+            }
+        };
+    }
+
+    /**
+     * @param int $intScale
+     * @param ImmutableDecimal $zeroTerm
+     * @return callable|ContinuedFractionTermInterface
+     */
+    protected function helperArcsecArccscBPart(
+        int $intScale,
+        ImmutableDecimal $zeroTerm
+    ): callable|ContinuedFractionTermInterface
+    {
+        /**
+         * @package Samsara\Fermat\Core
+         */
+        return new class($intScale, $zeroTerm) implements ContinuedFractionTermInterface {
+            private ImmutableDecimal $zeroTerm;
+            private int $scale;
+
+            /**
+             * @param int $scale
+             * @param ImmutableDecimal $zeroTerm
+             */
+            public function __construct(int $scale, ImmutableDecimal $zeroTerm)
+            {
+                $this->zeroTerm = $zeroTerm;
+                $this->scale = $scale;
+            }
+
+            /**
+             * @param int $n
+             * @return ImmutableDecimal
+             */
+            public function __invoke(int $n): ImmutableDecimal
+            {
+                if ($n == 0) {
+                    return $this->zeroTerm;
+                }
+
+                return new ImmutableDecimal(($n * 2) - 1, $this->scale);
+            }
+        };
     }
 
 }
